Had to change what accounts I am passing to the program. needed the user token accounts not the accounts of the tokens.
the test now passes, and 1 of each token is passed to the lp token accounts. I need to look into alternative ways to calulate the amount of the LP token.

There seems to be a formula standard based on the pro-rata share of the pool.

There also seems to be a special formula for the initializer which is the square of the two amounts. can implement that part first