Made quiet a few changes here.
We added a test to test the multiplication of large numbers because we end up with an overflow if the multiplied amounts are too large.

To handle this, we take the decimals into account, which means we must access the mint accounts to read the decimals of each token program, and handle them.

If we get an overflow error, we then account for this adjustment, rerun the calculation and then translate to the appropriate token amount of our lp token.

Unit tests now all passing